[Refactor] Entity 디렉토리를 Entities 디렉토리로 변경
- 대응되는 네임스페이스도 일괄 변경 처리
