+++ server/Controllers/UserController.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using server.DTOs;
-using server.Entity;
+using server.Entities;
 using server.Services;
